Use Math.abs() to ensure number is converted properly.
In some cases, the number is given to the method as a string.  When this
happens, the check for `isNan` and `!isFinite` passes if the string is formatted
like a number, but then number.toFixed() fails because that method doesn't exist
for strings.

By using Math.abs(), we're ensuring the number we're operating on is in fact
`typeof Number`.  It's also a bit clearer to read than the ternary statement
that was previously used, and we don't have to worry and fixing the polarity of
the number once we're done.